+++ flask_unchained/bundles/celery/extensions/__init__.py
@@ -1,0 +1,9 @@
+from .celery import Celery
+
+
+celery = Celery()
+
+
+EXTENSIONS = {
+    'celery': celery,
+}