+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * export Flask's `current_app` from the root package
 * never register static assets routes with babel bundle
 * integrate the mail bundle into the `flask_unchained` package
+* integrate the celery bundle into the `flask_unchained` package
 
 ## 0.4.2 (2018/07/21)